# refactor: num_matches to has_match