--- a/crates/subspace-proof-of-space/src/chiapos/tables.rs
+++ b/crates/subspace-proof-of-space/src/chiapos/tables.rs
@@ -7,7 +7,7 @@ extern crate alloc;
 use crate::chiapos::table::types::{Metadata, Position, X, Y};
 pub use crate::chiapos::table::TablesCache;
 use crate::chiapos::table::{
-    compute_f1, compute_fn, metadata_size_bytes, num_matches, partial_y, Table,
+    compute_f1, compute_fn, has_match, metadata_size_bytes, partial_y, Table,
     COMPUTE_F1_SIMD_FACTOR,
 };
 use crate::chiapos::utils::EvaluatableUsize;
@@ -375,14 +375,12 @@ where
         ys_and_metadata
             .array_chunks::<2>()
             .map(|&[(left_y, left_metadata), (right_y, right_metadata)]| {
-                (num_matches(left_y, right_y) == 1).then_some(compute_fn::<
+                has_match(left_y, right_y).then_some(compute_fn::<
                     K,
                     TABLE_NUMBER,
                     PARENT_TABLE_NUMBER,
                 >(
-                    left_y,
-                    left_metadata,
-                    right_metadata,
+                    left_y, left_metadata, right_metadata
                 ))
             })
             .collect()