@qy3u 2024-05-17
> I think it can make verification slightly faster on average.

Yes. The performance verification shows an improvement of approximately 15%.
before this commit:
```
chia/verification       time:   [17.590 µs 17.620 µs 17.650 µs]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
```
after:
```
chia/verification       time:   [14.867 µs 14.882 µs 14.896 µs]
                        change: [-15.695% -15.533% -15.343%] (p = 0.00 < 0.05)
                        Performance has improved.
```